Shopping List Item: delete. Edit, update, del  helper. Add progress bar

+++ app/src/main/java/com/example/shoppinglist/db/MainViewModel.kt
@@ -1,6 +1,7 @@
 package com.example.shoppinglist.db
 
 import androidx.lifecycle.*
+import com.example.shoppinglist.entities.LibraryItem
 import com.example.shoppinglist.entities.NoteItem
 import com.example.shoppinglist.entities.ShopListItem
 import com.example.shoppinglist.entities.ShoppingListNameItem
@@ -8,11 +9,15 @@ import kotlinx.coroutines.launch
 
 class MainViewModel(dataBase: MainDataBase): ViewModel() {
     val dao = dataBase.getDao()
+    val libraryItems = MutableLiveData<List<LibraryItem>>()
     val allNotes: LiveData<List<NoteItem>> = dao.getAllNotes().asLiveData()
     val allShopListNamesItem: LiveData<List<ShoppingListNameItem>> = dao.getAllShopListNames().asLiveData()
 
     fun getAllItemsFromList(listId: Int): LiveData<List<ShopListItem>>{
         return dao.getAllShopListItems(listId).asLiveData()
+    }
+    fun getAllLibraryItems(name: String)= viewModelScope.launch{
+        libraryItems.postValue(dao.getAllLibraryItems(name))
     }
 
     fun insertNote(note: NoteItem) = viewModelScope.launch {
@@ -21,9 +26,11 @@ class MainViewModel(dataBase: MainDataBase): ViewModel() {
 
     fun insertShopListName(listNameItem: ShoppingListNameItem) = viewModelScope.launch {
         dao.insertShopListName(listNameItem)
+
     }
     fun insertShopItem(shopListItem: ShopListItem) = viewModelScope.launch {
         dao.insertItem(shopListItem)
+        if(!isLibraryItemExist(shopListItem.name)) dao.insertLibraryItem(LibraryItem(null, shopListItem.name))
     }
 
     fun updateListItem(item: ShopListItem) = viewModelScope.launch {
@@ -38,12 +45,25 @@ class MainViewModel(dataBase: MainDataBase): ViewModel() {
         dao.updateListName(shopListNameItem)
     }
 
+    fun updateLibraryItem(item: LibraryItem) = viewModelScope.launch {
+        dao.updateLibraryItem(item)
+    }
+
     fun deleteNote(id: Int) = viewModelScope.launch {
         dao.deleteNote(id)
     }
 
-    fun deleteShoppingList(id: Int) = viewModelScope.launch {
-        dao.deleteShoppingList(id)
+    fun deleteLibraryItem(id: Int) = viewModelScope.launch {
+        dao.deleteLibraryItem(id)
+    }
+
+    fun deleteShopList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+        if(deleteList)dao.deleteShoppingList(id)
+        dao.deleteShopListItemsByListId(id)
+    }
+
+    private suspend fun isLibraryItemExist(name: String): Boolean{
+        return dao.getAllLibraryItems(name).isNotEmpty()
     }
 
     class  MainViewModelFactory(val database: MainDataBase) : ViewModelProvider.Factory{
